Use service account name also for cronjob pods
When pod security policies are used, then the securityContext typically
has a "runAsUser" attribute and a (non-default) service account.
The cronjob uses the same security context as the regular deployment, so
it should also use the same service account. With PSP enabled, a pod
created without this service account will not be able to start.

Signed-off-by: Aaron Spettl <aaron@spettl.de>