Only set REDIS_HOST_PASSWORD when auth is enabled
Due to the container's entrypoint, the configuration is messed up when
Redis is used without authentication and the environment variable
`REDIS_HOST_PASSWORD` is present (cf.
https://github.com/nextcloud/docker/blob/master/docker-entrypoint.sh#L122).

With the old setup, `REDIS_HOST_PASSWORD` will always be a string or an
empty string which is not enough. Now the chart only sets
`REDIS_HOST_PASSWORD` when `redis.auth.enabled` is `true`.

Signed-off-by: mjdazz <github@eleutheron.eu>