Update bitnami/postgresql to 12.12.10
This commit updates the postgresql dependency to version 12.12.10. This
is the latest 12.x version currently available, and while 13.x is
technically available, this would likely require bumping the nextcloud
chart version to 5.x, as PostgreSQL major version updates require manual
intervention in most cases (especially with the expected limited scope
of the embedded chart).

Signed-off-by: Archer <archer@nefarious.dev>