Include service in trusted domains for metrics
When nextcloud-exporter is used by enabling metrics, it fails to fetch
the serverinfo with 400 response code. This is related to the switch
from ingress to local service endpoint for nextcloud-exporter.
To fix this, in addition to .Values.nextcloud.host, the nextcloud
service also need to be included in the trusted domains.

Signed-off-by: Sunny <darkowlzz@protonmail.com>