Migrate the cron container from sidecar to K8s CronJob
As discussed in https://github.com/nextcloud/helm/issues/577

For me this has the major advantage that the CronJob can run as user
www-data (uid 33) and execute the php command directly whereas crond
in the sidecar has to run as root since otherwise the impersonation
of user www-data fails (crond: can't set groups: Operation not permitted).

Apart from that I think that is also more robust if the container fails
for some reason.

Signed-off-by: Martin Kirchner <martin.kirchner@cas.de>