Add dataVolumeMount option
By default, when persistence is enabled, the volume "nextcloud-main" is
used to store Nextcloud data. The chart offers an option to create an
extra PVC using persistence.nextcloudData.

However, it is not possible to mount data from an external source. Even
though extraVolumes and extraVolumeMounts are available, the data folder
is either sourced from "nextcloud-main" or "nextcloud-data".

By providing an option to skip mounting the "data" subfolder when using
the main persistence option, but not the nextcloudData option, it is
possible to provide a custom volume mount based on a different volume
specified in extraVolumes.

This allows, for example, to supply an NFS mount point:

```yaml
nextcloud:
  extraVolumes:
    - name: data
      nfs:
        path: /mnt/nextcloud
        server: my-nfs-server.example.com

  extraVolumeMounts:
    - name: data
      mountPath: /var/www/html/data

persistence:
  # Create a PVC to store Nextcloud application data
  enabled: true
  # But don't mount the user data folder, as it is provided above
  dataVolumeMount: false
```

Signed-off-by: Florian Ruechel <code@inexplicity.de>