Ensure version tag matches deployed version
When overriding `image.tag` the version label should also show the
correct version. Previously, the label would always be AppVersion, even
if the user had provided a different image tag (thus deploying a
different Nextcloud version).

With this change, the version is always first derived from the image tag
and only from AppVersion if not tag was provided. For the
"nextcloud.image" helper this was already the case, now the "version"
label on all resources also does this.

Signed-off-by: Florian Ruechel <code@inexplicity.de>